+++ test/test/capture_log_events_test.rb
@@ -35,5 +35,12 @@ class CaptureLogEventsTest < Minitest::Test
       assert_equal "User/disabled", capture_logger.events.last.metric
       assert capture_logger.events.last.metric_only?
     end
+
+    it "clears the events array when clear is called" do
+      capture_logger.info 'a message'
+      assert_equal(1, capture_logger.events.size)
+      capture_logger.clear
+      assert_equal(0, capture_logger.events.size)
+    end
   end
 end