+++ CHANGELOG.md
@@ -7,6 +7,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Move `newrelic_rpm.rb` mock to the `test/mocks` directory
 - Add `add_mocks_to_load_path`
 - Add NewRelicLogs formatter & appender
+- Move initialization of `events` array in `CaptureLogEvents` to constructor.
+- Add `clear` method to `CaptureLogEvents`.
 
 - Add payload logging to `payload_includes` Minitest helper functionality.
 - Fix bug in `payload_includes` testing functionality where if a key is not in the payload the test would pass.